+++ changelog.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.3] 2024-08-15
+
+### Changed
+
+- Added capability to add additional image content in the flexible-sections/greenville_image_content_module.php. This to be used as 'Featured'Listing. ACF Pro field groups where governing this and we do not have a license for the product to make changes to the field group. 
+- Created new flexible section for the Featured Listing content
+- Updated flexible content controller to use new module in place of previous.
+- Updated css to provide styling for new content. 
+
 ## [2.1.2] 2024-04-24
 
 ### Changed